modulize copilot logics
