+++ src/bcbench/agent/copilot/__init__.py
@@ -1,5 +1,5 @@
 """GitHub Copilot CLI agent module."""
 
-from bcbench.agent.copilot.copilot_agent import run_copilot_agent
+from bcbench.agent.copilot.agent import run_copilot_agent
 
 __all__ = ["run_copilot_agent"]
